requirements.txt
requests==2.31.0
python-dotenv==1.0.0
langchain==0.3.23
langchain-openai==0.3.14
pydantic>=2.7.4,<3.0.0
fastapi==0.115.9
uvicorn==0.34.2
parameterized==0.9.0
pytest==6.2.5
pytest-cov==4.1.0

